+++ server/middleware/getPhotoAndTextForHomepage.js
